Pulling changes from wip to GC-220-Inference

+++ DSL/Resql/global-classifier/POST/update-dataset-progress-session.sql
@@ -4,4 +4,4 @@ SET
     generation_message = :generation_message,
     progress_percentage = :progress_percentage,
     process_complete = :process_complete
-WHERE id = :id;
+WHERE id = :id::bigint;
